WIP on SetAcChargeTime
